Refactor: validation.js 파일로 유효성 검사 함수 분리
