+++ src/pages/auth/shared/lib/sign.js
@@ -1,13 +1,11 @@
-import { checkEmailDuplication, siginIn, signUp } from '../api/sign-api.js';
+import { signIn, signUp } from '../api/sign-api.js';
 import {
   EMAIL_ERROR_MSG_CLASSNAME,
   PASSWORD_CHECK_ERROR_MSG_CLASSNAME,
   PASSWORD_ERROR_MSG_CLASSNAME,
-  removeInputError,
   setInputError,
-  isValidEmail,
-  isValidPassword,
 } from './utils.js';
+import { validateEmailInput, validatePasswordCheckInput, validatePasswordInput } from './validation.js';
 
 if (localStorage.getItem('accessToken') !== null && localStorage.getItem('accessToken') !== 'undefined') {
   location.href = '/src/pages/folder';
@@ -29,48 +27,6 @@ const $passwordCheckBox = document.querySelector('.input-box-wrapper:has(#passwo
 const $passwordShowBtn = document.querySelector('.password-input-box .password-eye-controller');
 const $passwordCheckShowBtn = document.querySelector('.passwordcheck-input-box .password-eye-controller');
 
-const mockUserList = [
-  {
-    id: 'test@codeit.com',
-    ps: 'sprint101',
-  },
-];
-
-/**
- *
- * @param {HTMLInputElement} $email
- * @param {HTMLDivElement} $emailBox
- * @param {typeof EMAIL_ERROR_MSG_CLASSNAME} emailErrorClassName
- * @param {Event} event
- * @param {typeof formSignMode} formSignMode
- */
-const validateEmailInput = async ($email, $emailBox, emailErrorClassName, event, formSignMode) => {
-  if ($email.value === '') {
-    setInputError($email, '이메일을 입력해주세요.', emailErrorClassName, $emailBox);
-    return false;
-  }
-  if (!isValidEmail(reg => reg.test($email.value))) {
-    setInputError($email, '올바른 이메일 주소가 아닙니다.', emailErrorClassName, $emailBox);
-    return false;
-  }
-  if (formSignMode === 'signup' && (event.type === 'focusout' || event.type === 'submit')) {
-    const res = await checkEmailDuplication($email.value);
-    if (res.status === 409) {
-      setInputError($email, '이미 사용 중인 이메일입니다.', emailErrorClassName, $emailBox);
-      return false;
-    }
-  }
-  if (formSignMode === 'signin' && event.type === 'submit') {
-    const res = await checkEmailDuplication($email.value);
-    if (res.status !== 409) {
-      setInputError($email, '존재하지 않는 이메일입니다.', emailErrorClassName, $emailBox);
-      return false;
-    }
-  }
-  removeInputError({ $inputelem: $email, $errorMsgBox: document.querySelector(`.${emailErrorClassName}`) });
-  return true;
-};
-
 $email.addEventListener(
   'input',
   e => validateEmailInput($email, $emailBox, EMAIL_ERROR_MSG_CLASSNAME, e, formSignMode),
@@ -86,84 +42,20 @@ $email.addEventListener(
   },
 );
 
-/**
- *
- * @param {HTMLInputElement} $password
- * @param {HTMLDivElement} $passwordBox
- * @param {typeof PASSWORD_ERROR_MSG_CLASSNAME} passwordErrorClassName
- * @param {typeof formSignMode} formSignMode
- * @param {HTMLInputElement} $email
- * @returns
- */
-const validatePasswordInput = ($password, $passwordBox, passwordErrorClassName, formSignMode, $email) => {
-  if ($password.value === '') {
-    setInputError($password, '비밀번호를 입력해주세요.', passwordErrorClassName, $passwordBox);
-    return false;
-  }
-  if (!isValidPassword(reg => reg.test($password.value))) {
-    setInputError(
-      $password,
-      '비밀번호는 영문, 숫자 조합 8자 이상 입력해 주세요.',
-      passwordErrorClassName,
-      $passwordBox,
-    );
-    return false;
-  }
-  if (formSignMode === 'signin' && mockUserList.some(user => user.id === $email.value && user.ps !== $password.value)) {
-    setInputError($password, '비밀번호가 일치하지 않아요.', passwordErrorClassName, $passwordBox);
-    return false;
-  }
-  removeInputError({ $inputelem: $password, $errorMsgBox: document.querySelector(`.${passwordErrorClassName}`) });
-  return true;
-};
-
 $password.addEventListener(
   'input',
-  () => validatePasswordInput($password, $passwordBox, PASSWORD_ERROR_MSG_CLASSNAME, formSignMode, $email),
+  () => validatePasswordInput($password, $passwordBox, PASSWORD_ERROR_MSG_CLASSNAME),
   {
     signal,
   },
 );
 $password.addEventListener(
   'focusout',
-  () => validatePasswordInput($password, $passwordBox, PASSWORD_ERROR_MSG_CLASSNAME, formSignMode, $email),
+  () => validatePasswordInput($password, $passwordBox, PASSWORD_ERROR_MSG_CLASSNAME),
   {
     signal,
   },
 );
-
-/**
- *
- * @param {HTMLInputElement} $passwordCheck
- * @param {HTMLInputElement} $password
- * @param {HTMLDivElement} $passwordCheckBox
- * @param {typeof PASSWORD_CHECK_ERROR_MSG_CLASSNAME} passwordCheckErrorClassName
- * @returns
- */
-const validatePasswordCheckInput = ($passwordCheck, $password, $passwordCheckBox, passwordCheckErrorClassName) => {
-  if ($passwordCheck.value === '') {
-    setInputError($passwordCheck, '비밀번호를 입력해주세요.', passwordCheckErrorClassName, $passwordCheckBox);
-    return false;
-  }
-  if (!isValidPassword(reg => reg.test($passwordCheck.value))) {
-    setInputError(
-      $passwordCheck,
-      '비밀번호는 영문, 숫자 조합 8자 이상 입력해 주세요.',
-      passwordCheckErrorClassName,
-      $passwordCheckBox,
-    );
-    return false;
-  }
-  if ($passwordCheck.value !== $password.value) {
-    setInputError($passwordCheck, '비밀번호가 일치하지 않아요.', passwordCheckErrorClassName, $passwordCheckBox);
-    return false;
-  }
-  removeInputError({
-    $inputelem: $passwordCheck,
-    $errorMsgBox: document.querySelector(`.${passwordCheckErrorClassName}`),
-  });
-  return true;
-};
 
 if (formSignMode === 'signup') {
   $passwordCheck.addEventListener(
@@ -183,14 +75,8 @@ if (formSignMode === 'signup') {
  *
  * @param {SubmitEvent} e
  */
-const getValidationResults = async e => {
-  const pwValResult = !validatePasswordInput(
-    $password,
-    $passwordBox,
-    PASSWORD_ERROR_MSG_CLASSNAME,
-    formSignMode,
-    $email,
-  );
+const getValidationResultsOnSubmit = async e => {
+  const pwValResult = !validatePasswordInput($password, $passwordBox, PASSWORD_ERROR_MSG_CLASSNAME);
   const pwcheckValResult =
     formSignMode === 'signup' &&
     !validatePasswordCheckInput($passwordCheck, $password, $passwordCheckBox, PASSWORD_CHECK_ERROR_MSG_CLASSNAME);
@@ -204,14 +90,14 @@ const getValidationResults = async e => {
  * @param {HTMLInputElement} $email
  * @param {HTMLInputElement} $password
  */
-const getResponse = async (formSignMode, $email, $password) => {
+const getResponseOnSubmit = async (formSignMode, $email, $password) => {
   let res;
   switch (formSignMode) {
     case 'signup':
       res = await signUp({ email: $email.value, password: $password.value });
       break;
     case 'signin':
-      res = await siginIn({ email: $email.value, password: $password.value });
+      res = await signIn({ email: $email.value, password: $password.value });
       break;
     default:
       break;
@@ -221,22 +107,39 @@ const getResponse = async (formSignMode, $email, $password) => {
 
 /**
  *
+ * @param {HTMLFormElement} $form
+ */
+const resetInputs = $form => {
+  $form.querySelectorAll('input').forEach($input => {
+    $input.value = '';
+  });
+};
+
+/**
+ *
  * @param {SubmitEvent} e
  * @returns
  */
 const onSubmitHandler = async e => {
   e.preventDefault();
-  const [isInvalidEmail, isInvalidPassword, isInvalidPasswordCheck] = await getValidationResults(e);
+  const [isInvalidEmail, isInvalidPassword, isInvalidPasswordCheck] = await getValidationResultsOnSubmit(e);
   if (isInvalidEmail || isInvalidPassword || isInvalidPasswordCheck) {
     return;
   }
-  const res = await getResponse(formSignMode, $email, $password);
+  const res = await getResponseOnSubmit(formSignMode, $email, $password);
   if (res.ok) {
     const data = await res.json();
     const accessToken = await data.data.accessToken;
     localStorage.setItem('accessToken', JSON.stringify(accessToken));
+    resetInputs($form);
     location.href = '/src/pages/folder';
     controller.abort();
+  } else {
+    // 이메일 일치할 때 비밀번호 일치 여부 확인하려면 직접 signIn api 사용해봐야 알 수 있음.
+    // 어떤 이유로 실패했는지는 response에 담겨있지 않음.
+    if (res.status === 400 && formSignMode === 'signin') {
+      setInputError($password, '비밀번호가 일치하지 않아요.', PASSWORD_ERROR_MSG_CLASSNAME, $passwordBox);
+    }
   }
 };
 $form.addEventListener('submit', e => onSubmitHandler(e), { signal });
